+++ luna/src/core/scene/sceneSerializer.cpp
@@ -1,4 +1,5 @@
 #include "sceneSerializer.h"
+#include <core/scene/node.h>
 #include <core/vulkan/utils/vulkanAllocator.h>
 #include <yaml-cpp/yaml.h>
 #include <core/assets/assetManager.h>
