+++ luna/src/core/scripting/scriptUtils.cpp
@@ -38,5 +38,8 @@ namespace luna
 			void* param = &deltaTime;
 			if (m_ScriptClass->processMethod) mono_runtime_invoke(m_ScriptClass->processMethod, instance,&param, nullptr);
 		}
+		void scriptInstance::invokeSignal(std::string& signalName, void** params)
+		{
+		}
 	}
 }
